requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.5
python-dotenv==1.0.1
requests==2.32.3

langchain==0.2.17
langchain-community==0.2.17
langchain-openai==0.1.25
langchain-text-splitters==0.2.4

faiss-cpu==1.8.0.post1
pypdf==4.3.1
python-docx==1.1.2
sentence-transformers

streamlit==1.41.1

# optional for general web search tool:
tavily-python==0.5.0

